requirements.txt
torch
torchvision
numpy
Pillow
streamlit
scikit-learn
tqdm